PortScan.py aims to check open and closed ports.
PortScan.py aims to check open and closed ports quickly, It prompts you to inform a host that can be `LOCAL IP` `EXTERNAL IP` and `WEBSITE` because it converts the host informed and automatically searches its source IP address, it also asks to inform a start port and an end but if you want to check a only door you can too. it was developed in Python 3.5 and the System used was Arch Linux.